Feat(LottoDrawService): 랜덤 로또 구매하는 서비스 로직 구현

+++ src/main/kotlin/com/geniusjun/lotto/presentation/lotto/dto/LottoDrawResponse.kt
@@ -7,7 +7,7 @@ data class LottoDrawResponse(
     val bonusNumber: Int?,
     val bonusMatched: Boolean,
     val rank: String,
-    val reward: Int,
+    val reward: Long,
     val balance: Long
 ) {
     companion object {
@@ -18,7 +18,7 @@ data class LottoDrawResponse(
             bonusNumber: Int?,
             bonusMatched: Boolean,
             rank: String,
-            reward: Int,
+            reward: Long,
             balance: Long
         ) = LottoDrawResponse(
             myNumbers,
